test(http): extract mTLS e2e test into dedicated HttpMtlsTests (#842)
Move mtlsClientCertificate (and the HTTPS/needClientAuth WireMock
extension + PEM resource helpers) out of HttpTests into its own
HttpMtlsTests class in the same package.
